Ini Handler: Make more options case insensitive
There's a couple of options that use wcscmp() for no obvious reason -
switch these to _wcsicmp() to make them case insensitive to be
consistent with the rest of the ini parsing.

The remaining uses of wcscmp() are matching a default string "FLT_MAX"
that we provide so should not matter - and these should probably be
changed to use the newer GetIniFloat API instead.