Ini Handler: Merge IniSectionsMap and IniSectionsSorted
IniSectionsMap is an unsorted case insensitive mapping of sections to
keys, while IniSectionsSorted is a sorted case insensitive set of
section names to enable iterating over sections with a given prefix.

The section names in both of these data structures is identical, so
merge them into a single sorted case insensitive mapping of sections to
keys that can fulfil both of these uses simultaneously.