Ini Handler: Merge IniSectionVectorMap with IniSections
IniSections is a mapping of sections to a map of key/value pairs, and
IniSectionVectorMap is a mapping of sections to a vector of key/value
pairs.

Since the keys in both of these maps are identical we can merge them
into a single mapping of sections to both a map and a vector of
key/value pairs. Rather than adding a std::pair and using the
undescriptive "first" and "second" names, add a trivial struct to give
them readable names "kv_map" and "kv_vec".