Resource copying: Add a resource cache pool
Whenever we copy a resource by value we need to copy it to a resource
with a compatible description, which we cache, and if that same copy
operation later on needs to copy a resource with an incompatible
description it destroy that cache and create a new one with a compatible
description.

In most cases this is fine, because a single copy operation (or custom
resource target) will typically be used to copy the same resource over
and over, so we will rarely need to destroy and recreate the cache
(except for e.g. resolution changes). But there are some situations
where the cache for a single copy operation or custom resource may
legitimately be used to copy resources with differing descriptions.

e.g. One of the WATCH_DOGS2 updates introduced something like this where
one of the SSR shaders may be called on either a full resolution buffer
or a downscaled one, and both variants will be called each frame leading
us to continually destroy and recreate the cache (I'm not positive if
this is still the case in that game - this was an area of active
optimisation by Ubisoft).

This patch introduces a resource pool, so that multiple resources may be
held in the cache for each copy operation / custom resource target and
whichever one matches the current resource description will be used,
thereby saving us from having to destroy the old resources.

As a side effect of this change, we no longer need the 'Recreating
cached resource' log messages eliminating the log flood we could get if
a resource cache was being constantly destroyed and recreated. Instead
we log 'Switching cached resource', but we log that at debug level since
it does not indicate a problem.

The cache will also store NULL for any cached resources that failed to
be created - doing so allows us to skip retrying to create them over and
over again solving the log flood we previously had in this situation.

Note that the existing refcounting of the cached resource is left
unchanged and this just adds additional references to the resources in
the pool to stop them from being destroyed. This seemed to be the safest
option since the active resources can be copied by ref or value that go
through different code paths and messing with the existing refcounting
seemed like a sure-fire way to screw something up. We could potentially
drop the is_null optimisation though since the resource pool should also
stop the cached resources from being destroyed.

This commit is potentially a little risky for backwards compatibility -
adding the resource pool can slightly change the behaviour of 3DMigoto
in a potentially noticeable way in rare circumstances. e.g. before we
would get a new blank resource after recreating the buffer, but now we
will reuse an old one that potentially has data left on it - not likely
to be an issue if we are about to copy something over the top of it, but
if we were just using copy_desc this would be noticeable, but even then
unlikely that it would cause anyone problems.

I also think there might have been a rare edge case where previously we
could have places a reference to a resource from the game in a custom
resource, then performed a full copy from a compatible resource and
overwritten the resource we had a reference to, whereas now the code
will only allow a resource from the pool to be used when performing a
full copy, so will make a new cached resource with matching description.
The old behaviour wasn't really intended (I had the thought of adding an
"overwrite" keyword to allow this), and I don't think anyone is relying
on it so did not make any attempt to maintain compatibility with it. If
we do need it in the future (which is conceivable), I think it would be
better to make it explicit.