Add support for converting the mouse cursor into a texture
This is the major piece of the upcoming software mouse cursor support.
In addition to passing the mouse cursor X & Y screen coordinates into a
shader that we already had, this can also pass the hotspot X & Y
coordinates, and can convert the mouse cursor mask and color bitmaps
into DirectX textures and bind them to shader resource slots.