Add custom shaders to draw a software mouse cursor
This adds a custom shader that draws the mouse cursor in software to
allow the stereo depth to be adjusted.

It takes into account the cursor screen coordinates (which will need to
be fixed to use the window coordinates) and cursor hotspot offset to
position the icon on the screen, and runs two passes of the same custom
shader - the first to draw colour cursors to the screen, and the second
to draw black and white cursors to the screen using a different blend
mode that supports inverted cursors.

mousevs.hlsl can be edited to adjust the stereo depth of the cursor as
desired, possibly moving it to a fixed depth or doing something more
advanced like using crosshair.hlsl to line it up with the depth buffer.