Command list: Fix refcounting issues accessing built in resources
When accessing built in resources such as StereoParams, IniParams and
the cursor bitmaps we did not correctly increase the refcount, which
could lead to these being released prematurely sometimes causing a
crash - this was extremely evident when accessing the cursor resources
with the view cache enabled, probably because they are quickly destroyed
once the command list they are referenced from finishes execution.

This almost certainly is the correct fix for the crashes noted with
'vs-t125 = StereoParams', etc. and as such those are now uncommented in
the template d3dx.ini.

The view cache is re-enabled on the cursor resources (though since these
are still short lived it won't make much difference).