Resource copying: Use builtin views of StereoParams, IniParams, cursor bitmaps
StereoParams, IniParams and the cursor bitmaps counts as different types
to shader resources, so the resource copying code would not attempt to
use the view that we had already constructed for these and instead
construct (and cache) its own view of these resources, which works but
is wasteful.

This adds an equivalence function that reports these builtin resources
as being compatible with 'shader resources' so that if they are assigned
to a SRV slot the builtin view will be used directly in much the same
way as assigning one SRV directly to another by reference.