Software Mouse: Correct colours
The cursor bitmaps are BGRA, not RGBA - so switch to the appropriate
format to switch the channels the right way around. Also switch to an
sRGB format to shade it correctly, and tweak the alpha blending to match
the real cursor.

Animated cursors still don't work, but the next step is to hide the
hardware cursor.