Software Mouse: Add hide_cursor option to hide the hardware mouse cursor
This turned out to be stupidly complicated.

One might expect that simply calling ShowCursor(FALSE); would be
sufficient to hide the hardware mouse cursor, especially since it has a
counter to deal with the cursor being hidden and shown from multiple
locations... but, although the call appears to succeed it doesn't
actually alter the cursor visibility depending on where (which thread?)
it was called from (and the same goes for SetCursor()).

So, we hook into the ShowCursor() function so that when it is called we
will be running in the context of the game thread where it will work. We
maintain our own cursor visibility counter separate from Windows, and
from this call will synchronise the hardware counter to ours, possibly
hiding the cursor if requested in the d3dx.ini.

But it doesn't stop there - since we have hidden the hardware cursor
GetCursorInfo() won't return it and our software mouse implementation
will fail. So we also hook into SetCursor() to catch the cursor handles
from the game, and hook GetCursorInfo() so we can return them when the
software cursor is supposed to be visible, but the hardware cursor is
not.