Software Mouse: Clarify show stopper Steam Bug
Returning the hardware cursor visibility count from ShowCursor() was
necessary to work around a bug in the Steam overlay that inconsistently
uses the hooked vs trampoline version of ShowCursor (bypasses the hooks
and gets the hardware visibility counter when opening the overlay, fails
to bypass the hooks and sets the software visibility counter when
closing the overlay), but unfortunately it turns out that doing that
messes up other games - e.g. Witcher 3 hangs at a black screen on the
main menu, so we really do need to return the software visibility
counter and either do something about the Steam overlay, or try a
different strategy that will work with either.