Software mouse: Second attempt to suppress hardware cursor
This is a second attempt to suppress the hardware cursor that will
hopefully work for all games & overlays - this time we leave the
visibility count alone and instead replace the cursor icon with a
completely invisible one. Since the hardware cursor technically is
displayed, the visibility counts for software and hardware cursors
match, so we no longer need to manage them separately. We hook into
SetCursor, GetCursor and GetCursorInfo to keep a handle of the cursor
the game set and return it whenever something (including our own
software mouse implementation) asks for it.