Software Mouse: Add cursor_window_x/y to get window coordinates
cursor_screen_x/y were a first pass at adding mouse support to 3DMigoto
ages back, but were only ever intended to be used in fullscreen games
where the screen coordinates match the window coordinates. With a
generic software mouse implementation it becomes more important to try
to get this right even in cases where the game may not be running in
full screen.

This adds cursor_window_x / cursor_window_y to the available ini param
overrides in the command list, which will call ScreenToClient() to
translate the cursor screen coordinates into window coordinates.

cursor_x and cursor_y are shorthand for these, as I expect they will be
overall more useful than cursor_screen_x/y (which we can probably
consider deprecating if this works well).

This does not yet obtain the hWnd necessary to actually do that
translation, which will come in a subsequent patch to minimise the
impact of potential merge conflicts with the upscaling code.