Software Mouse: Use render target instead of screen resolution + cleanup
If the game is windowed and smaller than the resolution, the software
cursor would be scaled and positioned incorrectly as it was basing this
off the screen resolution from the driver. Pass the resolution of the
active render target (which will be the back buffer) to the shader, and
use that to calculate the correct position and scale regardless of the
window size.

Also clean up the vertex shader a little to use #defines to declare
which IniParams are which at the top of the shader, making the shader
easier to read, and easier to tweak if the parameters need to be
changed.