Ini Handler: Fix dying if [Profile] section is missing
There is a critical BUG exit in GetIniSection, as it expected to only be
called on sections that had already been enumerated elsewhere and
therefore should exist... except we also call it on the [Profile]
section explicitly, which is an optional section and is valid to not
exist.

Downgrade the BUG to a WARNING and return an empty data structure if the
section doesn't exist.