Software Cursor: Hook DefWindowProc to fix cursor in some games
Some games such as Akiba's Trip do not directly call SetCursor() and
instead rely on the WM_SETCURSOR message to set the cursor, and these
were not handled by our software cursor implementation - the hardware
cursor would not be suppressed and the software cursor could display the
incorrect cursor (e.g. busy instead of arrow).

This is due to the fact that the WM_SETCURSOR processing in
DefWindowProc() is able to directly set the cursor without calling
the exported SetCursor() function that we have hooked.

To solve this, we hook DefWindowProcA() and DefWindowProcW() and emulate
the behaviour of the original procedure (including passing the message
to the parent window), but calling into our hooked version of
SetCursor() whenever the mouse cursor is in the client area of the
window and a valid window class cursor is found.

For cases where the cursor is outside of the client area we actually do
want the real hardware cursor to be displayed, so we call back into the
original DefWindowProc() to handle that, intentionally bypassing our
SetCursor() hook and coping with any other edge cases there may be.