Software Mouse: (Hopefully) fix rendering colour cursors with no alpha channel
Some cursors have an alpha channel (separate from the AND mask) and some
don't, and Windows does not seem to provide any API to determine which
is in use, and the bit depth does not indicate this because 32bpp can
either mean 24bpp + 8bit alpha, or 24bpp + 8bit unused. People on
stackoverflow are scanning the entire alpha channel looking for non-zero
values to fudge this.

The problem is that from the shader if we read the alpha channel as 0 it
could mean that either there is no alpha channel, or there is and this
pixel should be fully transparent. Since the mask is valid in either
case we treat alpha=0 to mean no alpha channel and set the pixel to
fully opaque, hoping that the mask will then clear it if it was supposed
to be transparent.

This might fall down if someone makes a cursor with a blank mask that
relies entirely on the alpha channel.