Software Cursor: Add support for animated cursors
The bitmaps that GetIconInfo() gives us only include the first frame of
an animated cursor, and it would be nice to do better. The DrawIconEx()
API allows us to specify which frame we are drawing, so we use that to
draw a frame from the animated cursor to another bitmap, then convert
that bitmap into a DirectX texture. This API accepts a flag which can
allow us to use it for both the mask and colour of a colour cursor, and
will also work for static cursors, but will not work for black and white
cursors since their mask bitmap is double height - for those we skip
this call and use the old procedure.

To determine which frame we want we have to turn to an undocumented
Windows API GetCursorFrameInfo(), which scraping together information
from stackoverflow, WINE, and various resources on animated cursors
returns the number of frames and period of the animated cursor in
multiples of 1/60 seconds. We use that information along with the global
tick count to calculate which frame to display - that way we don't have
to track how long since the cursor was last assigned.