Give cursor_x/y a different meaning to cursor_window_x/y
The software mouse shader needs the cursor position in window
coordinates since it does actually want to work in pixels, but in most
cases where we pass the cursor position into a game we really want to
have it in the range [0:1] to sample a texture or [-1:1] to match up
with clip space.

Rather than always require the user pass in the window size and divide
it in the shader, let's make cursor_x/y scale it to [0:1] and save
cursor_window/screen_x/y for situations where we need it to be in
pixels.