d3dx.ini: Use a [CommandList] to simplify unbinding all render targets
All three of our custom shaders unbind all render and depth targets,
which is a bit verbose. This adds a [CommandListUnbindAllRenderTargets]
that can be used as a shorthand in these custom shaders and from
anywhere else people might want to do the same.

It should be treated the same as adding each of these lines into the
sections individually, so it is worth keeping in mind that if any of
these need to be copied (or copy_desc, etc) that should be done before
running the command list, and binding your own render targets should be
done after calling this.