stereo2mono: Use fallback path when stereo is disabled
If stereo is disabled the reverse stereo blit won't work and we would
end up with the destination only updated on the left, which may lead to
shaders reading stale or 0 data if they read from the right hand side.

Repro case:
1. Copy separation from StereoParams.Load(0).x into a stereo Texture2D
2. Use stereo2mono to turn it into a double width mono Texture2D
3. Read out the values from each side of the texture with stereo on and
   note they are correct
4. Disable stereo via Ctrl+T
5. Repeat steps 1 and 2
6. Read out the values from each side of the resource with stereo on and
   note that one side is zero since stereo is disabled, but the other
   side still contains the old value of separation from before stereo
   was disabled

This patch enables the fallback path in the stereo2mono copy for cases
where stereo is disabled, so that both sides of the destination texture
will contain an identical copy of the mono side of the texture being
copied.