d3dx.ini: Add 'post' keyword to examples in present and document
The post keyword in the present section marks something to be run at the
start of a frame instead of the end. In a lot of cases it won't make a
practical difference, but it is a good habbit to clear resources at the
start of a frame instead of the end, just in case you add a debug shader
at the end of a frame that uses them without remembering that they may
have been cleared depending on the order in the section. Clearing them
at the start of a frame means you probably won't have to worry.