Custom Shaders: Allow depth/stencil state to be overridden
This allows us to override more state from the game for our own
purposes, and it's fully flexible so the modder can set the state to
whatever they need.

One example would be to trigger a custom shader section from an
transparency pass that enables write access to the depth buffer and
uses a replacement depth buffer, allowing us to capture depth
information from transparent objects for use with automatic HUD
adjustments.