Frame Analysis: Ignore 'log' option and always create a frame analysis log
The frame analysis log has proven extremely useful and doesn't slow down
frame analysis by any significant margin, so always enable it regardless
of whether it was requested or not. The option is still left in so the
parser doesn't complain, but has no effect anymore.

In particular, this avoids accidentally disabling the log for part of
the frame if a persist trigger has been set up that doesn't mention log.