Frame Analysis & Debug logs: Fix excess newline before resource hash
The following commit added newlines to some calls that intentionally had
them missing, which then broke up these lines in the frame analysis
logs, or added extra newlines that shouldn't have been there. One of
these had also been inserted before a close bracket, resulting in a
rogue ')' on the next line in the log:

    commit c16f932c22f5fff7e8c802b19626f3768b538a0d
    Author: bo3b <github@bo3b.net>
    Date:   Sat May 27 11:58:13 2017 -0700

        Once more, with completeness.  Fix all missing \n.

The ones that were intentionally missing were immediately followed by a
call to FrameAnalysisLogResourceHash(), which appends a resource hash to
a line in the frame analysis log followed by a newline, or just inserts
a newline if the resource is NULL. There was a legitimate problem fixed
by the above commit, since the newline was not also inserted into the
debug log.

This commit removes the newlines on every line that is immediately
followed by that call, and adds a DebugLog("\n") into
FrameAnalysisLogResourceHash() so that it will complete the line in the
debug log as well. For our reference, this adds a 'NoNL' to the end of
the FrameAnalysisLog calls, to indicate that the lack of a newline is
done intentionally. For now that makes no difference to calling the
function without it and is just for our reference, but later if we
refactor the logging it may help distinguish between the calls.