AfterDraw: Don't skip post draw processing if handling=skip
Handling=skip will not only skip the draw call, but will also skip some
of the processing we perform afterwards. This can be a problem because a
lot of that processing is undoing state changes that we might have made
beforehand regardless of the skip setting (e.g. separation=0;
handling=skip will set the separation, but never restore it).

It also prevents executing the post command list despite the pre command
list being executed, and prevents frame analysis from running for that
draw call (which may seem superfluous, but consider that handling=skip
may be used alongside run=CustomShaderBlah, or otherwise calling one of
the draw calls from the command list. e.g. I use this on HUD shaders
that have complex #includes to force them to always be recompiled).

I can't see a good reason to skip any of this processing, so remove the
fast exit from AfterDraw() so we always perform the processing.