Shader Overrides: Make it very obvious we want to deprecate all the old filters
The command list is already capable of replacing many of the old legacy
filters, and we will be better off if we continue moving these into the
command list. Add a large if (ENABLE_LEGACY_FILTERS) around the old ones
to very clearly indicate that we don't want to continue adding to them.

The if will be optimised out by the compiler since ENABLE_LEGACY_FILTERS
is defined to 1 - this is purely there for the benefit of anyone looking
at the code, though in the future we may move that into the d3dx.ini
for a few versions before we decide to actually remove any of these.