Presets: Accept both 'preset = Foo' and 'preset = PresetFoo'
The original preset code did not accept the "Preset" prefix on the
prefix command, as in it would only accept 'preset = Foo', not 'preset =
PresetFoo'. While I agree that the later is redundant since the word
preset now appears twice, it is more consistent with the way we have
referenced other sections in the command list (ps-t0 = ResourceBar, run
= CustomShaderBaz, etc), and it makes it easier to search for
'PresetFoo' to find both where it is used and where it is referenced, so
it is good to support here... but for backwards compatibility and less
redundancy for those that prefer not to say "preset" twice we support
both ways.