Presets: Use OverrideGlobalSave instead of custom save & restore
The following does not work correctly:

    [Constants]
    ; Inventory depth, see also: PresetInventory
    x2 = 3.4

    [ShaderOverrideHUDSettingsAndInventory]
    hash = bab4f74d620346c8
    ...
    ; Check for the inventory texture to activate a low separation preset:
    checktextureoverride = ps-t0

    [TextureOverrideInventoryExamineCircles]
    hash = b98c5309
    filter_index = 7
    preset = PresetExamineInventory

    [PresetExamineInventory]
    convergence = 0.5
    x2 = 0.475
    transition = 400
    transition_type = cosine
    release_transition = 300
    release_transition_type = cosine

When the preset is activated x2 correctly transitions to 0.475, however
when the preset is deactivated it does not return to the original value
of 3.4. Rather than invent yet another save & restore mechanism, let's
reuse the global save & restore object that was already created for key
overrides and handles this situation. If there are any problems with it
that make it unsuitable for presets, we should look to address those -
for instance, refer to some of the brainstorming in:
https://github.com/bo3b/3Dmigoto/issues/33