Presets: Allow multiple presets to be activated simultaneously
The preset code only allowed a single preset to be active
simultaneously, which precludes possibilities such as having multiple
presets with independent triggers operating on independent variables. It
also restricted the active preset to be the last preset triggered during
a frame.

This commit relaxes that restriction, allowing multiple presets to be
activated and deactivated independently just as key bindings are. If
only one trigger is present in a frame this will be equivalent to the
previous behaviour, and if multiple presets affect the same variable
than the value from the most recently activated preset will be used
so long as it remains active.

There are some possible improvements we might want to make, such as
forbidding a deactivated preset from returning to a value set by a
different preset that has since been deactivated (the global undo area
will prevent these from having any permanent changes), and assigning
priorities to each preset to dictate if any should be considered more
important than others in the event that they both set the same value.