Resource Copying: Fix crash when copying a buffer with unknown stride
If we copy a constant buffer we take StructureByteStride in the buffer
description to be the stride, but this can be zero. If we then copy the
buffer into any slot which takes a view we use this stride when
calculating FirstElement and NumElements, but if the stride was zero we
will crash with a divide by zero error.

Check for zero stride and take it to mean that there is only a single
element.