Software Mouse: Render software mouse before upscaling
This switches around the order of the upscaling and software mouse
shaders so that the software mouse behaves correctly when both are in
use.

I'm not convinced that this will work in all cases - I think what this
actually means is that we are no longer emulating the hardware mouse
exactly, but it just so happens to fix issues that the hardware mouse
had with upscaling, at least in some games.

Tested in Witcher 3... but then Witcher 3 locks the hardware mouse to a
small region that matches what it thinks it set the resolution to (but
thanks to upscaling is not), so moving the software mouse shader here
will effectively counter that, but what if the game behaved differently
- what if it did not lock the mouse and asked Windows what the window
size was? Would this change now break games with that behaviour?

I suspect the correct fix might actually be to intercept a lot of
windowing and mouse calls, translating coordinates between Windows and
the game, effectively fixing the hardware cursor when upscaling is
enabled, and once that works the software mouse shader should be
returned to where it was.

But for now, this is easier and will do until we have tested the
behaviour in a wider range of games with upscaling enabled to find out
whether my nagging feeling is justified or not. Worst case the user can
always flip these sections around since it is just a d3dx.ini change.