Resource copying: Fix copying constant buffers into view slots
If we attempt to copy a constant buffer into a slot that requires a view
(Unordered access view, shader resource, render target) we currently
fail, but it is the view creation that fails, not the resource creation.
This can actually somewhat mask the fact that it failed (aside from all
the errors in the log) as if the resource was copied the [Resource]
object will still hold a valid buffer with valid information and if we
copy that back into a constant buffer slot it will appear to have
worked, but if we try to access it from another slot type it will not.

This is especially a problem if trying to use the constant buffer as a
UAV to update some values before re-injecting it back into the original
shader, as it will appear that the values were not changed. It is more
obvious when copying using copy_desc, since then the UAV version won't
have valid data which will show up when copying it back into the
original.

Constant buffers have no format, no stride, are not structured buffers
and have no view. In order to create a view for a buffer we need to fill
in either a valid format, or a stride. Since constant buffers may
contain mixed types and are accessed in a similar manner as structured
buffers, opt to fill in the stride value, and with no better value to
use, fill it in as the total buffer size.