Command lists: Fix NULL pointer dereference in draw = from_caller
If draw = from_caller is used in a command list that was not called as a
result of a draw call (e.g. from a dispatch call or present) we would
dereference a NULL pointer and crash. While it's tempting to say "don't
do that", it's better not to crash, so check for a NULL pointer and bail
out safely.