Software Mouse and Upscaling Fix-First Version
Added code to support software mouse together with upscaling.
It should work with all games that uses the GetCursorPos to read the current mouse pos.
However there are other possibilities than that (eg. WM_MOUSEMOVE - bad but who knows what the game engines do)
Please note the comments in the DLLMainHook.cpp (GetCursorInfo and ScreenToClient) and in the CommandList.cpp (UpdateCursorInfo)
It is also possible at least in this version to revert changes from commit 337aab0 (This commit does not do it)