Upscaling / SetCursorPos: Fix scaling coordinates in wrong direction
When the game calls SetMousePos we need to translate the coordinates
from the game resolution back to the real resolution - swap around the
multiply and divide to correct this. Tested in Dreamfall Chapters when
clicking on an item with multiple interaction items the game calls this
to set the mouse to the middle of the available choices.