Hooking: Use helper function to reduce boiler plate
The amount of boiler plate code we needed to install a new hook was
quite excessive, and takes the focus away from what is actually
important in each of the functions. Refactor the various hook install
routines to use a helper function that takes care of the boiler plate
around each individual hook so we only have one line of code for each
new hook, and minimal boiler plate in the function as a whole.

This also gets rid of the old structs we used to hold the hook ID of the
LoadLibraryExW since we never actually use the hook ID.