Upscaling / mouse: Hook GetClientRect to return original width & height
This fixes the behaviour of the software mouse cursor since it now will
scale itself based on what the game thinks the resolution is set to
instead of the real resolution.

The software mouse cursor is now scaled up, because it will render the
cursor texture 1:1 on the fake back buffer before upscaling has taken
place. We would need to have it bypass our hooks to allow otherwise, and
we should probably make that call before the next release as it will be
a pain to swap later.