Upscaling / Mouse: Fix ScreenToClient translation
GetClientRect() rather counter-intuitively returns a rectangle that is
relative to the client area itself, not the screen, so top and left will
always be 0. We therefore do need to call the original ScreenToClient()
to translate the coordinates from the screen to client area instead of
doing it ourselves, but we need to handle the scaling that we apply.

We assume that we will have already given the game the faked coordinates
earlier and so the coordinates it passes in here will be fake - first,
scale them back to the real screen coordinates before calling
ScreenToClient(), then after that call scale them back to the fake
coordinates before returning them to the game.