Software Cursor: Add an undocumented cursor_upscaling_bypass option
This option is mostly for our internal testing, and it gives us a way to
specify whether the mouse should use the same translation that the
upscaling code does for any cursor, or whether it should bypass it and
use the real cursor position and window size.

If all is working properly*, this should only alter the size of the
cursor - with this set to false the cursor will be subject to upscaling,
and with this set to true it will not be.

This can be used in conjunction with rendering the software cursor
before or after the upscaling shader to eliminate pixelation - to match
the appearance of the hardware cursor exactly (assuming there are no
other issues*) it should be set to true and the software cursor rendered
after upscaling. Doing the opposite on both still gives a nice
appearance, but the cursor is oversized. Mismatching the two options
will give a pixelated cursor.

* issues such as in Witcher 3 with upscaling=1 instead of upscaling=2,
  which seems to use a different hwnd for their fullscreen window.