Hooking: Cleanup trampoline function pointers as suggested by Bo3b
This implements the suggestion by Bo3b from:
https://github.com/bo3b/3Dmigoto/commit/312f21e6b906437983ba47b2b2c9e979bc990745#commitcomment-24059866

It eliminates most of the typedefs, instead specifying the type directly
in the trampoline function pointer itself. The only exception I left in
was for DefWindowProcA/W, since both variants have the same function
prototype.

This also sets the function pointers to default to the original
functions where they did not already do so, so that any code that wants
to call the original function bypassing our hooks can do so without
having to check if the pointer has been initialised first.

This also removes the Hooked_LoadLibraryExW and Hooked_IsDebuggerPresent
function prototypes, which are not needed since the file is already in
the right order so that they are implemented before they are referenced.