Upscaling / Mouse: Always use up to date window size for scaling
We don't want to use what resolution the user requested in the d3dx.ini
when translating the mouse coordinates, because it may not match
reality. Always get the current window size using the trampoline
GetClientRect function to perform the translation. Fixes issue with
hardware cursor location in Witcher 3 running in windowed upscaling 720p
to 720p but requested upscaling to 1080p.