d3dx.ini: Note that upscaling=2 might fix issues with the mouse cursor
Witcher 3 has some odd behaviour with their hwnd in full screen mode
(not the case in windowed mode), where the hwnd they pass DirectX does
not seem to be the same one they use for their fullscreen window, nor
the one they use to check the window size for mouse cursor calculations.

With upscaling=1 this leads to problem where the window they hand us is
still the original non-upscaled resolution, the game clips the hardware
cursor to it, and the hardware cursor does not line up with the menu
items (the software cursor *does* line up with the menu, but this is a
fluke since the software cursor is *supposed* to line up with the
hardware cursor).

upscaling=2 works around both of these issues, since we force the window
they pass DirectX to fullscreen and therefore the desired resolution.
Since this may impact other games, add a note in the d3dx.ini that
upscaling=2 might help with mouse issues.