Software Cursor: Add linear filters
If upscaling and the software cursor are enabled but there is either a
mismatch between rendering the cursor before / after upscaling and
bypassing the upscaling cursor translations, or there is some other
unresolved issue we might get a pixelated cursor, which looks a little
nasty. Implement a linear filter in the mouse pixel shader so that if we
run into this situation (but things are otherwise working) the mouse
cursor will look acceptable. This helps in Witcher3 with upscaling=1, or
when cursor_upscaling_bypass is disabled.

Because of the way we have to handle the cursor mask and colour textures
we can't simply use a sampler to interpolate it, so this adds a very
simple linear filter using Loads().

If the cursor is being drawn 1:1 this looks identical (just as sharp) as
before, but if the cursor is being scaled for whatever reason this now
looks much nicer.