Resource copying: Implement raw views
This adds a "raw" keyword to resource copying to facilitate raw views
into buffers when used as a UAV or SRV, e.g.

    [ResourceByteAddressable]
    type = ByteAddressBuffer
    format = R32_TYPELESS
    array = 50

    [CustomShaderRawCS]
    cs-u0 = raw ResourceByteAddressable
    post cs-u0 = null

    [CustomShaderRawPS]
    ps-t100 = raw ResourceByteAddressable
    post ps-t100 = null

Note that I've observed that at least some buffers created to for raw
access may not be accessed in other ways, which may mean that we may be
able to drop or imply the raw keyword in those cases, but I'm not
convinced that they are always mutually exclusive - further
investigation required.