Resource Copying: Implement keywords to copy and reset the UAV counter
The counter associated with a UAV cannot be directly accessed, and must
be copied to another resource to read, and set while binding the UAV to
the pipeline.

Add a 'reset_counter' keyword that can be used to reset the UAV counter
while binding it to UAV slot on the pipeline.

Add a 'copy_counter' keyword to arbitrary resource copying to indicate
that it should be the counter that is copied instead of the original
resource. Note that as I understand it the counter is associated with a
view, not a resource - meaning that trying to copy the counter from a
CustomResource will fail since we don't currently store the UAV in with
the CustomResource.  Instead, the counter should be copied from a view
bound to the pipeline, like:

    [ResourceAppendable]
    type = AppendStructuredBuffer
    stride = 32
    array = 50

    [ResourceCounter]

    [CustomShaderFoo]
    cs-u0 = reset_counter append ResourceAppendable
    post ResourceCounter = copy_counter cs-u0
    post cs-u0 = null

The biggest limitation here is that since the counter is associated with
a view, and since the views are currently cached with each copy
operation we can't reset it from the present call. It could be useful
today to capture a bunch of information out of a single invocation (e.g.
record all vertices or pixels that match some filter), but not across
multiple invocations. Moving the view caches into the CustomResource
being bound to the pipeline would give us more flexibility here.

Note for our eventual port to DX12 that it has changed how the counter
is stored.