Add depth_stencil_state_merge option
This option changes the behaviour of the depth stencil state override in
the custom shader sections.

If it is set to false, the previous behaviour of entirely replacing the
depth/stencil state with default values + overrides is retained (if the
state is overridden at all), and the state object will be created once
when the custom shader is substantiated.

If it is set to true then each time the custom shader is used it will
merge any overridden state in with the state from the game. This is
useful to e.g. override the depth_write mask without touching
depth_enable.