Support index & depth buffer filtering + checktextureoverride in the command list
This extends the texture filtering and checktextureoverride support in
the command list to work with any of the resource target types that is
supported by the arbitrary resource copying support.

This deprecates (as if it wasn't already) indexbufferfilter, as that can
now be achieved with e.g. 'x = ib' and processed in the shader.