Upscaling: Fix rendering ARK in a corner
ARK: Survival Evolved (UE4) only renders in a corner when upscaling is
enabled. It seems that during the ResizeTarget() call, DirectX itself
must call GetClientRect(), and because we return the faked client area
it uses that for the size of the window.

We want DirectX to get the real resolution, so temporarily disable the
translation in the GetClientRect() call around the original
ResizeTarget() call.