Command List: Refactor to avoid passing device & context everywhere
It is getting a little too verbose passing the hacker and original
device and contexts to every call that needs them in the command list
code. Place pointers to these objects in the command list state object,
and pass that around instead - that way we only have one pointer to pass
around and a lot of functions already had access to this object.