Add rasterizer_state_merge option
This is the rasterizer state equivalent of depth_stencil_state_merge,
and allows the rasterizer state to be partially overridden.

This is largely for completeness, though it is worth noting that the
rasterizer state override could already have been used to selectively
disable or enable scissor clipping since way back in February of last
year for 3DMigoto 1.2.32, and this makes it easier since we no longer
have to reconstruct the rest of the state.

It doesn't reuse the merged state object though (if not merging it
does), so in that regard disable_scissor is still superior to this, but
disable_scissor lacks the ability to override other rasterizer state
that this provides (e.g. I've used this in the past to change certain
shaders to render in wireframe for debugging purposes, or to disable
culling to fix a certain blue light glitch in WATCH_DOGS2).